test: split drawings Postgres suites by concern
