+++ test/integration/drawings-postgres.test.ts
@@ -5,13 +5,15 @@ import { readFile } from 'node:fs/promises'
 import { setTimeout as delay } from 'node:timers/promises'
 import test, { mock } from 'node:test'
 import { Client } from 'pg'
-import { Hono, type Context } from 'hono'
-import { NETWORK, USDC } from '../../src/chain.ts'
-import { beginCityCreditSpend, issueCityFeeCredit } from '../../src/city-credit.ts'
-import type { Resident } from '../../src/core.ts'
-import { bindPaymentEvidence, canonicalPaymentRequest } from '../../src/payment-attempts.ts'
-import { completeTreasuryPaymentOperation } from '../../src/payment-treasury-operations.ts'
-import { TREASURY } from '../../src/pay.ts'
+import { founderWorldDrawing } from '../helpers/drawings-fixtures/drawing-samples.ts'
+import { registerSchemaContractTests } from './drawings-tests/schema-contract.ts'
+import { registerWorldMigrationTests } from './drawings-tests/world-migrations.ts'
+import { registerDrawingConstraintsTests } from './drawings-tests/drawing-constraints.ts'
+import { registerSnapshotInheritanceTests } from './drawings-tests/snapshot-inheritance.ts'
+import { registerDrawingRoutesTests } from './drawings-tests/drawing-routes.ts'
+import { registerPaidRevisionTests } from './drawings-tests/paid-revisions.ts'
+import { registerLateFinalityTests } from './drawings-tests/late-finality.ts'
+import { registerLegacyPaymentTests } from './drawings-tests/legacy-payments.ts'
 
 const POSTGRES_IMAGE =
   'postgres@sha256:7958605b474b3d264a969cb3a123d6aa00ad1e1fe9da8a69984dabb704d93317'
@@ -75,26 +77,8 @@ const drawing = (colour: string) => ({
   indices: [0, ...blankIndices().slice(1)],
 })
 const blankDrawing = () => ({ palette: [], indices: blankIndices() })
-const rowsFor = (value: { indices: readonly (number | null)[] }): string[] =>
-  Array.from({ length: 8 }, (_, row) => value.indices
-    .slice(row * 8, row * 8 + 8)
-    .map(index => index == null ? '.' : String(index))
-    .join(' '))
 const ownerSecret = `1f3d9_sk_${'d'.repeat(48)}`
 const ownerAuthorization = Object.freeze({ authorization: `Bearer ${ownerSecret}` })
-const founderWorldDrawing = Object.freeze({
-  palette: Object.freeze(['#0b1714', '#123026', '#1c4434']),
-  indices: Object.freeze([
-    0, 0, 0, 0, 0, 0, 0, 0,
-    null, 0, 1, 0, 0, 0, 0, 0,
-    null, 0, 0, 0, 0, 0, 1, 0,
-    0, null, 0, 0, 1, 0, 0, 0,
-    0, 0, 0, 0, 0, 0, 0, 0,
-    1, 0, null, 0, 1, 0, 0, 0,
-    0, 0, null, 0, 0, 0, 0, 1,
-    0, 0, 0, 0, 0, 1, 0, 0,
-  ]),
-})
 
 function docker(args: readonly string[], allowFailure = false): string {
   const result = spawnSync('docker', [...args], {
@@ -191,53 +175,9 @@ test('real PostgreSQL enforces, preserves, moderates, exports, and settles drawi
   })
   routeClient = client
 
-  await client.query('BEGIN')
-  try {
-    await client.query('CREATE SCHEMA drawing_contract_probe')
-    await client.query('SET LOCAL search_path TO drawing_contract_probe, public')
-    for (const name of [
-      'valid_city_drawing',
-      'valid_city_drawing_public_text',
-      'valid_city_drawing_variant_name',
-      'valid_city_drawing_state',
-      'valid_city_drawing_variants',
-      'city_drawing_rows',
-      'city_drawing_presentation_state',
-      'valid_city_drawing_revision_value',
-      'city_drawing_public_value',
-    ]) await client.query(drawingFunctionDdl(name))
-
-    await client.query('SET LOCAL search_path TO pg_catalog')
-    const customSchemaContract = (await client.query<{
-      safe_state: boolean
-      unsafe_state: boolean
-      trimmed_variant: boolean
-      public_state: string
-    }>(`
-      SELECT
-        drawing_contract_probe.valid_city_drawing_state(
-          'refused', 'Owner chose not to draw this.', NULL
-        ) AS safe_state,
-        drawing_contract_probe.valid_city_drawing_state(
-          'refused', $1, NULL
-        ) AS unsafe_state,
-        drawing_contract_probe.valid_city_drawing_variant_name(
-          ' padded variant '
-        ) AS trimmed_variant,
-        drawing_contract_probe.city_drawing_public_value(
-          'refused', 'Owner chose not to draw this.', NULL,
-          'thing', NULL, NULL, NULL, NULL
-        )->>'presentation_state' AS public_state
-    `, ['hidden\u202Elabel'])).rows[0]
-    assert.deepEqual(customSchemaContract, {
-      safe_state: true,
-      unsafe_state: false,
-      trimmed_variant: false,
-      public_state: 'refused',
-    })
-  } finally {
-    await client.query('ROLLBACK')
-  }
+  await registerSchemaContractTests(
+    t, client, drawingFunctionDdl,
+  )
 
   await client.query(schemaDdl)
   const freshWorld = (await client.query<{ id: number; drawing: unknown }>(`
@@ -245,126 +185,11 @@ test('real PostgreSQL enforces, preserves, moderates, exports, and settles drawi
   `)).rows[0]!
   assert.deepEqual(freshWorld.drawing, founderWorldDrawing)
 
-  await client.query(migrationDdl)
-  await client.query(migrationDdl)
-  await installLegacyNullWorldDrawing(client)
-  await client.query(worldRootDrawingMigrationDdl)
-  await client.query(worldRootDrawingMigrationDdl)
-  await client.query(migrationDdl)
-  await client.query(worldRootTopologyMigrationDdl)
-  await client.query(worldRootTopologyMigrationDdl)
-  await client.query(drawingContractMigrationDdl)
-  await client.query(drawingContractMigrationDdl)
-
-  await client.query(`
-    ALTER TABLE places DROP CONSTRAINT places_drawing_contract;
-    ALTER TABLE places ALTER COLUMN drawing_state DROP NOT NULL;
-    ALTER TABLE places ALTER COLUMN drawing_state DROP DEFAULT;
-    ALTER TABLE places DISABLE TRIGGER places_protect_topology_write;
-    UPDATE places SET drawing_state = NULL, drawing_description = NULL
-    WHERE place_kind = 'world';
-    ALTER TABLE places ENABLE TRIGGER places_protect_topology_write;
-
-    CREATE FUNCTION reject_unrelated_place_backfill() RETURNS trigger
-    LANGUAGE plpgsql AS $function$
-    BEGIN
-      RAISE EXCEPTION 'unrelated protected place must not receive a row update';
-    END
-    $function$;
-    CREATE TRIGGER places_reject_unrelated_backfill
-    BEFORE UPDATE ON places
-    FOR EACH ROW EXECUTE FUNCTION reject_unrelated_place_backfill();
-  `)
-  try {
-    await client.query(drawingContractMigrationDdl)
-    assert.deepEqual((await client.query<{ drawing_state: string; tgenabled: string }>(`
-      SELECT place.drawing_state, trigger.tgenabled
-      FROM places place
-      JOIN pg_trigger trigger ON trigger.tgrelid = 'places'::regclass
-        AND trigger.tgname = 'places_reject_unrelated_backfill'
-        AND NOT trigger.tgisinternal
-      WHERE place.place_kind = 'world'
-    `)).rows, [{ drawing_state: 'complete', tgenabled: 'O' }])
-  } finally {
-    await client.query(`
-      DROP TRIGGER IF EXISTS places_reject_unrelated_backfill ON places;
-      DROP FUNCTION IF EXISTS reject_unrelated_place_backfill();
-    `)
-  }
-
-  const migratedWorld = (await client.query<{ id: number; drawing: unknown }>(`
-    SELECT id, drawing FROM places WHERE place_kind = 'world'
-  `)).rows[0]!
-  assert.deepEqual(migratedWorld, freshWorld)
-  assert.deepEqual((await client.query<{ conname: string; convalidated: boolean }>(`
-    SELECT conname, convalidated FROM pg_constraint
-    WHERE conrelid = 'places'::regclass
-      AND conname IN ('places_world_shape', 'places_world_drawing_exact')
-    ORDER BY conname
-  `)).rows, [
-    { conname: 'places_world_drawing_exact', convalidated: true },
-    { conname: 'places_world_shape', convalidated: true },
-  ])
-  assert.deepEqual((await client.query<{ tgenabled: string }>(`
-    SELECT tgenabled FROM pg_trigger
-    WHERE tgrelid = 'places'::regclass
-      AND tgname = 'places_protect_topology_write'
-      AND NOT tgisinternal
-  `)).rows, [{ tgenabled: 'O' }])
-
-  await client.query('BEGIN')
-  try {
-    await client.query('ALTER TABLE places DISABLE TRIGGER places_protect_topology_write')
-    await assert.rejects(
-      client.query(
-        "UPDATE places SET drawing = $1::jsonb WHERE place_kind = 'world'",
-        [JSON.stringify(drawing('#174d3c'))],
-      ),
-      (error: unknown) => (error as { code?: string }).code === '23514',
-    )
-  } finally {
-    await client.query('ROLLBACK')
-  }
-
-  await installLegacyNullWorldDrawing(client)
-  await client.query(schemaDdl)
-  await client.query(schemaDdl)
-  assert.deepEqual((await client.query<{ id: number; drawing: unknown }>(`
-    SELECT id, drawing FROM places WHERE place_kind = 'world'
-  `)).rows[0], freshWorld)
-
-  const columns = (await client.query<{ table_name: string; column_name: string }>(`
-    SELECT table_name, column_name
-    FROM information_schema.columns
-    WHERE table_schema = 'public'
-      AND (
-        (table_name IN ('residents', 'places', 'things', 'kind_revisions')
-          AND column_name IN ('drawing', 'drawing_state', 'drawing_description'))
-        OR (table_name = 'kind_revisions' AND column_name = 'drawing_variants')
-        OR (table_name = 'things' AND column_name = 'drawing_variant_name')
-      )
-    ORDER BY table_name, column_name
-  `)).rows
-  assert.deepEqual(columns, [
-    { table_name: 'kind_revisions', column_name: 'drawing' },
-    { table_name: 'kind_revisions', column_name: 'drawing_description' },
-    { table_name: 'kind_revisions', column_name: 'drawing_state' },
-    { table_name: 'kind_revisions', column_name: 'drawing_variants' },
-    { table_name: 'places', column_name: 'drawing' },
-    { table_name: 'places', column_name: 'drawing_description' },
-    { table_name: 'places', column_name: 'drawing_state' },
-    { table_name: 'residents', column_name: 'drawing' },
-    { table_name: 'residents', column_name: 'drawing_description' },
-    { table_name: 'residents', column_name: 'drawing_state' },
-    { table_name: 'things', column_name: 'drawing' },
-    { table_name: 'things', column_name: 'drawing_description' },
-    { table_name: 'things', column_name: 'drawing_state' },
-    { table_name: 'things', column_name: 'drawing_variant_name' },
-  ])
-  assert.equal((await client.query<{ count: string }>(`
-    SELECT count(*)::text AS count FROM information_schema.tables
-    WHERE table_schema = 'public' AND table_name = 'drawing_revisions'
-  `)).rows[0]?.count, '1')
+  await registerWorldMigrationTests(
+    t, client, freshWorld, migrationDdl, installLegacyNullWorldDrawing,
+    worldRootDrawingMigrationDdl, worldRootTopologyMigrationDdl,
+    drawingContractMigrationDdl, schemaDdl, drawing,
+  )
 
   const residentDrawing = drawing('#ad3f25')
   const placeDrawing = drawing('#174d3c')
@@ -447,67 +272,8 @@ test('real PostgreSQL enforces, preserves, moderates, exports, and settles drawi
   ]) await rejectsCheck(client, candidate)
 
   const replacement = drawing('#0b1714')
-  await client.query(`
-    UPDATE residents
-    SET drawing = $1::jsonb,
-      drawing_state = 'in_progress',
-      drawing_description = 'The replacement is deliberately unfinished.'
-    WHERE id = 1
-  `, [JSON.stringify(replacement)])
-  assert.deepEqual(
-    (await client.query<{
-      drawing: unknown
-      drawing_state: string
-      drawing_description: string
-    }>(`
-      SELECT drawing, drawing_state, drawing_description FROM residents WHERE id = 1
-    `)).rows[0],
-    {
-      drawing: replacement,
-      drawing_state: 'in_progress',
-      drawing_description: 'The replacement is deliberately unfinished.',
-    },
-  )
-
-  await assert.rejects(
-    client.query(`
-      UPDATE things
-      SET drawing = $2::jsonb,
-        drawing_state = 'complete',
-        drawing_description = 'Typed things may not invent instance pixels.'
-      WHERE id = $1
-    `, [inheritedThingId, JSON.stringify(thingDrawing)]),
-    (error: unknown) => (error as { code?: string }).code === '23514',
-  )
-  await assert.rejects(
-    client.query(`
-      INSERT INTO kind_revisions (
-        kind_id, revision, description, traits, recipe,
-        drawing, drawing_state, drawing_description, drawing_variants
-      )
-      SELECT kind_id, 2, description, traits, recipe,
-        drawing, drawing_state, drawing_description, $2::jsonb
-      FROM kind_revisions
-      WHERE kind_id = $1 AND revision = 1
-    `, [kindId, JSON.stringify(Array.from({ length: 9 }, (_, index) => ({
-      name: `variant-${index}`,
-      drawing: thingDrawing,
-      state: 'complete',
-      description: `Variant ${index}.`,
-    })))]),
-    (error: unknown) => (error as { code?: string }).code === '23514',
-  )
-
-  await assert.rejects(
-    client.query(
-      "UPDATE places SET drawing = $1::jsonb WHERE place_kind = 'world'",
-      [JSON.stringify(placeDrawing)],
-    ),
-    (error: unknown) => (error as { code?: string }).code === '55000',
-  )
-  assert.deepEqual(
-    (await client.query<{ drawing: unknown }>("SELECT drawing FROM places WHERE place_kind = 'world'")).rows[0]?.drawing,
-    founderWorldDrawing,
+  await registerDrawingConstraintsTests(
+    t, client, replacement, inheritedThingId, thingDrawing, kindId, placeDrawing,
   )
 
   await client.query(`
@@ -522,579 +288,29 @@ test('real PostgreSQL enforces, preserves, moderates, exports, and settles drawi
     `, [className, recordId])
   ).rows[0]!.payload
 
-  assert.deepEqual((await snapshot('residents', 1)).drawing, replacement)
-  assert.equal((await snapshot('residents', 1)).drawing_state, 'in_progress')
-  assert.equal(
-    (await snapshot('residents', 1)).drawing_description,
-    'The replacement is deliberately unfinished.',
+  await registerSnapshotInheritanceTests(
+    t, client, snapshot, replacement, freshWorld, placeId, placeDrawing,
+    kindId, kindDrawing, ownThingId, thingDrawing, inheritedThingId,
+    variantThingId, kindVariants,
   )
-  assert.deepEqual((await snapshot('residents', 1)).drawing_rows, rowsFor(replacement))
-  assert.equal((await snapshot('residents', 1)).drawing_source, 'resident')
-  assert.deepEqual((await snapshot('places', freshWorld.id)).drawing, founderWorldDrawing)
-  assert.equal((await snapshot('places', freshWorld.id)).drawing_state, 'complete')
-  assert.deepEqual((await snapshot('places', freshWorld.id)).drawing_rows, rowsFor(founderWorldDrawing))
-  assert.equal((await snapshot('places', freshWorld.id)).drawing_source, 'place')
-  assert.deepEqual((await snapshot('places', placeId)).drawing, placeDrawing)
-  assert.equal((await snapshot('places', placeId)).drawing_source, 'place')
-  assert.deepEqual((await snapshot('kinds', kindId)).drawing, kindDrawing)
-  assert.equal((await snapshot('kinds', kindId)).drawing_source, 'kind_base')
-  assert.equal((await snapshot('kinds', kindId)).revision, 1)
-  assert.deepEqual((await snapshot('things', ownThingId)).drawing, thingDrawing)
-  assert.equal((await snapshot('things', ownThingId)).drawing_source, 'thing')
-  assert.deepEqual((await snapshot('things', inheritedThingId)).drawing, kindDrawing)
-  assert.equal((await snapshot('things', inheritedThingId)).drawing_source, 'kind_base')
-  assert.equal((await snapshot('things', inheritedThingId)).kind_id, kindId)
-  assert.equal((await snapshot('things', inheritedThingId)).revision, 1)
-  assert.deepEqual((await snapshot('things', variantThingId)).drawing, kindVariants[0]!.drawing)
-  assert.equal((await snapshot('things', variantThingId)).drawing_source, 'kind_variant')
-  assert.equal((await snapshot('things', variantThingId)).variant_name, 'blue-shutter')
-
-  await client.query('UPDATE things SET owner_id = 2 WHERE id = $1', [variantThingId])
-  assert.deepEqual((await snapshot('things', variantThingId)).drawing, kindVariants[0]!.drawing)
-  assert.equal((await snapshot('things', variantThingId)).drawing_source, 'kind_variant')
-  assert.equal((await snapshot('things', variantThingId)).variant_name, 'blue-shutter')
-
-  await client.query(`
-    UPDATE things
-    SET drawing = NULL,
-      drawing_state = 'refused',
-      drawing_description = 'The thing owner refuses the selected kind variant.'
-    WHERE id = $1
-  `, [variantThingId])
-  const refusedVariantSnapshot = await snapshot('things', variantThingId)
-  assert.equal(refusedVariantSnapshot.drawing_state, 'refused')
-  assert.equal(refusedVariantSnapshot.drawing_source, 'thing')
-  assert.equal(refusedVariantSnapshot.variant_name, null)
-
-  const { mountDrawingRoutes: mountRefusalDrawingRoutes } = await import('../../src/drawings.ts')
-  const refusalDrawingApp = new Hono()
-  mountRefusalDrawingRoutes(refusalDrawingApp, {
-    database: {
-      query: async (text: string, params: readonly unknown[] = []) =>
-        (await client!.query(text, [...params])).rows,
-    },
-    authenticate: async () => null,
-  })
-  const refusedVariantCurrent = await refusalDrawingApp.request(
-    `/api/drawing/thing/${variantThingId}`,
-  )
-  assert.equal(refusedVariantCurrent.status, 200)
-  assert.deepEqual(await refusedVariantCurrent.json(), {
-    type: 'thing', id: variantThingId,
-    state: 'refused', presentation_state: 'refused',
-    description: 'The thing owner refuses the selected kind variant.',
-    drawing: null, rows: null, source: 'thing',
-    kind_id: kindId, revision: 1,
-  })
-  await client.query(`
-    UPDATE things
-    SET drawing = NULL, drawing_state = 'undrawn', drawing_description = NULL
-    WHERE id = $1
-  `, [variantThingId])
-  assert.equal((await snapshot('things', variantThingId)).drawing_source, 'kind_variant')
-  assert.equal((await snapshot('things', variantThingId)).variant_name, 'blue-shutter')
-
-  await client.query(`
-    UPDATE things
-    SET drawing = NULL,
-      drawing_state = 'refused',
-      drawing_description = 'The thing owner refuses inherited artwork.'
-    WHERE id = $1
-  `, [inheritedThingId])
-  assert.equal((await snapshot('things', inheritedThingId)).drawing, null)
-  assert.equal((await snapshot('things', inheritedThingId)).drawing_state, 'refused')
-  assert.equal((await snapshot('things', inheritedThingId)).drawing_source, 'thing')
-  await client.query(`
-    UPDATE things
-    SET drawing = NULL, drawing_state = 'undrawn', drawing_description = NULL
-    WHERE id = $1
-  `, [inheritedThingId])
-  assert.deepEqual((await snapshot('things', inheritedThingId)).drawing, kindDrawing)
-  assert.equal((await snapshot('things', inheritedThingId)).drawing_source, 'kind_base')
-  const eventId = Number((await client.query<{ id: number }>(`
-    SELECT id FROM events WHERE kind = 'resident_edited' ORDER BY id DESC LIMIT 1
-  `)).rows[0]!.id)
-  assert.deepEqual((await snapshot('events', eventId)).detail, {
-    resident_id: 1, source_thing_id: ownThingId,
-  })
-
-  await client.query(`
-    INSERT INTO drawing_revisions (
-      target_type, target_id,
-      prior_state, prior_description, prior_drawing, prior_source,
-      prior_kind_id, prior_kind_revision,
-      current_state, current_description, current_drawing, current_source,
-      current_kind_id, current_kind_revision, current_variant_name,
-      slot_variant_name, author_id, author_relation
-    ) VALUES (
-      'thing', $1,
-      'complete', 'The kind owner’s plain lantern.', $2::jsonb, 'kind_base',
-      $3, 1,
-      'complete', 'A blue shutter over the kind owner’s lantern.', $4::jsonb, 'kind_variant',
-      $3, 1, 'blue-shutter',
-      'blue-shutter', 1, 'owner'
-    )
-  `, [
-    inheritedThingId,
-    JSON.stringify(kindDrawing),
-    kindId,
-    JSON.stringify(kindVariants[0]!.drawing),
-  ])
-
-  await client.query(`
-    INSERT INTO moderation_actions (target_type, target_id, action, actor_id, reason)
-    VALUES
-      ('resident', 1, 'remove', 1, 'hide resident drawing'),
-      ('kind', $1, 'remove', 1, 'hide kind drawing')
-  `, [kindId])
-  const hiddenResident = await snapshot('residents', 1)
-  assert.equal(hiddenResident.handle, 'drawing-owner')
-  assert.equal(hiddenResident.drawing, null)
-  assert.equal((await snapshot('kinds', kindId)).status, 'maintainer_hidden')
-  assert.equal((await snapshot('things', inheritedThingId)).drawing, null)
-  assert.equal((await snapshot('things', inheritedThingId)).drawing_source, null)
-  assert.deepEqual((await snapshot('things', ownThingId)).drawing, thingDrawing)
 
   const database = {
     query: async (text: string, params: readonly unknown[] = []) =>
       (await client!.query(text, [...params])).rows,
   }
-  const drawingOwner: Resident = Object.freeze({
-    id: 2,
-    handle: 'paid-drawing-owner',
-    model: 'integration',
-    joined_at: '2026-08-27T00:00:00.000Z',
-    quota_day: '2026-08-27',
-    things_today: 0,
-    notes_today: 0,
-    agreement_actions_today: 0,
-  })
-  const { mountDrawingRoutes } = await import('../../src/drawings.ts')
-  const drawingApp = new Hono()
-  mountDrawingRoutes(drawingApp, {
-    database,
-    authenticate: async (_context: Context) => drawingOwner,
-  })
-  const worldDrawingRead = await drawingApp.request(`/api/drawing/place/${freshWorld.id}`)
-  assert.equal(worldDrawingRead.status, 200)
-  assert.deepEqual(await worldDrawingRead.json(), {
-    type: 'place',
-    id: freshWorld.id,
-    state: 'complete',
-    presentation_state: 'complete',
-    description: '',
-    drawing: founderWorldDrawing,
-    rows: rowsFor(founderWorldDrawing),
-    source: 'place',
-  })
-  const checkpoint = (await client.query<{ checkpoint: string }>(`
-    SELECT current_change_id::text AS checkpoint
-    FROM public_change_state
-    WHERE singleton = true
-  `)).rows[0]!.checkpoint
-  const worldThumbnailRead = await drawingApp.request(
-    `/api/drawing/place/${freshWorld.id}/thumb.png?rev=${checkpoint}`,
-  )
-  assert.equal(worldThumbnailRead.status, 200)
-  assert.equal(worldThumbnailRead.headers.get('content-type'), 'image/png')
-  assert.equal(
-    worldThumbnailRead.headers.get('cache-control'),
-    'public, max-age=31536000, immutable',
-  )
-  const worldThumbnailBytes = new Uint8Array(await worldThumbnailRead.arrayBuffer())
-  assert.deepEqual([...worldThumbnailBytes.slice(0, 8)], [137, 80, 78, 71, 13, 10, 26, 10])
-  const hiddenInheritedCurrent = await drawingApp.request(
-    `/api/drawing/thing/${inheritedThingId}`,
-  )
-  assert.equal(hiddenInheritedCurrent.status, 200)
-  assert.deepEqual(await hiddenInheritedCurrent.json(), {
-    type: 'thing', id: inheritedThingId,
-    state: 'undrawn', presentation_state: 'undrawn',
-    description: null, drawing: null, rows: null, source: 'none',
-  })
-  const hiddenInheritedThumbnail = await drawingApp.request(
-    `/api/drawing/thing/${inheritedThingId}/thumb.png?rev=${checkpoint}`,
-  )
-  assert.equal(hiddenInheritedThumbnail.status, 404)
-  assert.equal(hiddenInheritedThumbnail.headers.get('cache-control'), 'no-store')
-  const hiddenInheritedHistory = await drawingApp.request(
-    `/api/drawing/thing/${inheritedThingId}/history`,
-  )
-  assert.equal(hiddenInheritedHistory.status, 200)
-  assert.deepEqual(await hiddenInheritedHistory.json(), {
-    type: 'thing', id: inheritedThingId,
-    revisions: [],
-    page: { limit: 20, has_more: false, next_before: null },
-  })
-  const admittedColours = ['#101010', '#202020', '#303030', '#404040', '#505050', '#606060']
-  for (const colour of admittedColours) {
-    const response = await drawingApp.request('/api/me/drawing', {
-      method: 'PATCH',
-      body: JSON.stringify({
-        drawing: drawing(colour),
-        drawing_state: 'complete',
-        drawing_description: 'A resident rate-limit proof drawing.',
-      }),
-    })
-    assert.equal(response.status, 200, await response.clone().text())
-    assert.equal((await response.json() as { changed: boolean }).changed, true)
-  }
-  const limited = await drawingApp.request('/api/me/drawing', {
-    method: 'PATCH',
-    body: JSON.stringify({
-      drawing: drawing('#707070'),
-      drawing_state: 'complete',
-      drawing_description: 'A resident rate-limit proof drawing.',
-    }),
-  })
-  assert.equal(limited.status, 429)
-  assert.equal(limited.headers.get('retry-after'), '60')
-  const exactRetry = await drawingApp.request('/api/me/drawing', {
-    method: 'PATCH',
-    body: JSON.stringify({
-      drawing: drawing(admittedColours.at(-1)!),
-      drawing_state: 'complete',
-      drawing_description: 'A resident rate-limit proof drawing.',
-    }),
-  })
-  assert.equal(exactRetry.status, 200)
-  assert.equal((await exactRetry.json() as { changed: boolean }).changed, false)
-  assert.deepEqual((await client.query<{ used: number }>(`
-    SELECT used FROM resident_drawing_rate_limits WHERE resident_id = 2
-  `)).rows.map(row => Number(row.used)), [6])
-  assert.equal(Number((await client.query<{ count: string }>(`
-    SELECT count(*)::text AS count FROM events
-    WHERE kind = 'resident_edited' AND actor = 'paid-drawing-owner'
-  `)).rows[0]?.count), 6)
-
-  const firstPaidDrawing = drawing('#8a622d')
-  await issueCityFeeCredit(database, {
-    founderId: 1,
-    residentId: 2,
-    sourceKey: 'drawing-kind-invention-credit-0001',
-    reason: 'real PostgreSQL drawing invention proof',
-  })
-  const invention = await beginCityCreditSpend(database, {
-    actorId: 2,
-    operation: 'kind_invention',
-    targetKey: 'kind-invention:paid-drawn-kind',
-    requestId: 'drawing-kind-invention-request-0001',
-    request: {
-      name: 'paid-drawn-kind',
-      description: 'a paid kind with a pinned drawing revision',
-      traits: [],
-      recipe: [],
-      drawing: firstPaidDrawing,
-      drawing_state: 'complete',
-      drawing_description: 'The first paid kind drawing.',
-    },
-  })
-  assert.equal(invention.state, 'ready')
-  if (invention.state !== 'ready') assert.fail('expected a ready paid kind invention')
-  const invented = await completeTreasuryPaymentOperation(database, {
-    attemptId: invention.attempt_id,
-    leaseOwner: invention.lease_owner,
-  })
-  assert.equal(invented.state, 'completed')
-  if (invented.state !== 'completed') assert.fail('expected a completed paid kind invention')
-  assert.equal(invented.status, 201)
-  const inventedKind = invented.response.kind as { id: number; drawing: unknown }
-  assert.deepEqual(inventedKind.drawing, firstPaidDrawing)
-
-  const pinnedPaidThingId = Number((await client.query<{ id: number }>(`
-    INSERT INTO things (
-      place_id, name, body, owner_id, maker_id, kind_id,
-      birth_revision, current_revision
-    ) VALUES ($1, 'paid pinned thing', '', 2, 2, $2, 1, 1)
-    RETURNING id
-  `, [placeId, inventedKind.id])).rows[0]!.id)
-  const secondPaidDrawing = drawing('#1e5964')
-  await issueCityFeeCredit(database, {
-    founderId: 1,
-    residentId: 2,
-    sourceKey: 'drawing-kind-revision-credit-0001',
-    reason: 'real PostgreSQL drawing revision proof',
-  })
-  const revision = await beginCityCreditSpend(database, {
-    actorId: 2,
-    operation: 'kind_revision',
-    targetKey: `kind-revision:${inventedKind.id}:2`,
-    requestId: 'drawing-kind-revision-request-0001',
-    assetType: 'kind',
-    assetId: inventedKind.id,
-    request: {
-      kind_id: inventedKind.id,
-      description: 'the second paid drawing revision',
-      traits: [],
-      recipe: [],
-      drawing: secondPaidDrawing,
-      drawing_state: 'complete',
-      drawing_description: 'The second paid kind drawing.',
-    },
-  })
-  assert.equal(revision.state, 'ready')
-  if (revision.state !== 'ready') assert.fail('expected a ready paid kind revision')
-  const revised = await completeTreasuryPaymentOperation(database, {
-    attemptId: revision.attempt_id,
-    leaseOwner: revision.lease_owner,
-  })
-  assert.equal(revised.state, 'completed')
-  if (revised.state !== 'completed') assert.fail('expected a completed paid kind revision')
-  assert.equal(revised.status, 200)
-  assert.deepEqual((revised.response.kind as { drawing: unknown }).drawing, secondPaidDrawing)
-
-  const pinnedDrawing = async () => (await client!.query<{ drawing: unknown }>(`
-    SELECT revision.drawing
-    FROM things thing
-    JOIN kind_revisions revision
-      ON revision.kind_id = thing.kind_id AND revision.revision = thing.current_revision
-    WHERE thing.id = $1
-  `, [pinnedPaidThingId])).rows[0]!.drawing
-  assert.deepEqual(await pinnedDrawing(), firstPaidDrawing)
-  await client.query('UPDATE things SET current_revision = 2 WHERE id = $1', [pinnedPaidThingId])
-  assert.deepEqual(await pinnedDrawing(), secondPaidDrawing)
-
-  const lateFinalityDrawing = drawing('#6f4b7d')
-  const lateFinalityRequest = {
-    name: 'late-finality-drawn-kind',
-    description: 'a drawing-bearing x402 invention finalized after its authorization window',
-    traits: [],
-    recipe: [],
-    drawing: lateFinalityDrawing,
-    drawing_state: 'complete',
-    drawing_description: 'The late-finality kind drawing.',
-  }
-  const lateFinalityCanonical = canonicalPaymentRequest(lateFinalityRequest)
-  const lateFinalityAttemptId = 'drawing-late-finality-attempt-0001'
-  const lateFinalityLeaseOwner = 'drawing-late-finality-lease-0001'
-  const lateFinalityTxHash = `0x${'7'.repeat(64)}`
-  const lateFinalityPayer = `0x${'5'.repeat(40)}`
-  const lateFinalityResponseHeader = Buffer.from(JSON.stringify({
-    success: true,
-    transaction: lateFinalityTxHash,
-    payer: lateFinalityPayer,
-    network: NETWORK,
-  })).toString('base64')
-  const lateBoundary = (await client.query<{
-    start_time: Date
-    end_time: Date
-    block_time: Date
-    valid_after: string
-    valid_before: string
-  }>(`
-    WITH observed AS MATERIALIZED (SELECT clock_timestamp() AS at)
-    SELECT at AS start_time,
-      at + interval '2 seconds' AS end_time,
-      at + interval '1 second' AS block_time,
-      extract(epoch FROM date_trunc('second', at))::bigint - 1 AS valid_after,
-      extract(epoch FROM date_trunc('second', at))::bigint + 2 AS valid_before
-    FROM observed
-  `)).rows[0]!
-  assert.ok(lateBoundary.block_time >= lateBoundary.start_time)
-  assert.ok(lateBoundary.block_time < lateBoundary.end_time)
-  await client.query(`
-    INSERT INTO payment_attempts (
-      public_id, actor_id, operation, target_key, request_hash, request_json,
-      method, network, token, payer_wallet, payee_wallet, amount_units,
-      x402_nonce, x402_payload_digest, x402_valid_after, x402_valid_before,
-      start_block, start_time, end_time, status, lease_owner, lease_expires_at
-    ) VALUES (
-      $1, 2, 'kind_invention', $2, $3, $4::jsonb,
-      'x402', $5, $6, $7, $8, 1000000,
-      $9, repeat('a', 64), $10::bigint, $11::bigint,
-      50000000, $12::timestamptz, $13::timestamptz,
-      'settling', $14, clock_timestamp() + interval '30 seconds'
-    )
-  `, [
-    lateFinalityAttemptId,
-    `kind-invention:${lateFinalityRequest.name}`,
-    lateFinalityCanonical.hash,
-    lateFinalityCanonical.json,
-    NETWORK,
-    USDC.toLowerCase(),
-    lateFinalityPayer,
-    TREASURY,
-    `0x${'6'.repeat(64)}`,
-    lateBoundary.valid_after,
-    lateBoundary.valid_before,
-    lateBoundary.start_time.toISOString(),
-    lateBoundary.end_time.toISOString(),
-    lateFinalityLeaseOwner,
-  ])
-
-  const pendingLateFinality = await bindPaymentEvidence(database, {
-    publicId: lateFinalityAttemptId,
-    leaseOwner: lateFinalityLeaseOwner,
-    txHash: lateFinalityTxHash,
-    finality: null,
-    paymentResponseHeader: lateFinalityResponseHeader,
-  })
-  assert.equal(pendingLateFinality.status, 'payment_pending')
-  assert.ok(pendingLateFinality.recoveryDeadlineAt)
-
-  const wallClockStartedAt = Date.now()
-  let finalizedAt = ''
-  for (;;) {
-    const observed = (await client.query<{
-      observed_at: Date
-      operation_passed: boolean
-      authorization_passed: boolean
-    }>(`
-      SELECT clock_timestamp() AS observed_at,
-        clock_timestamp() > $1::timestamptz + interval '1 millisecond' AS operation_passed,
-        clock_timestamp() > to_timestamp($2::bigint) + interval '1 millisecond'
-          AS authorization_passed
-    `, [lateBoundary.end_time.toISOString(), lateBoundary.valid_before])).rows[0]!
-    if (observed.operation_passed && observed.authorization_passed) {
-      finalizedAt = observed.observed_at.toISOString()
-      break
-    }
-    await delay(25)
-  }
-  assert.ok(
-    Date.now() - wallClockStartedAt >= 1_500,
-    'the drawing test must cross a real PostgreSQL authorization window',
+  await registerDrawingRoutesTests(
+    t, client, database, freshWorld, inheritedThingId, drawing,
   )
 
-  const finalizedLatePayment = await bindPaymentEvidence(database, {
-    publicId: lateFinalityAttemptId,
-    leaseOwner: lateFinalityLeaseOwner,
-    txHash: lateFinalityTxHash,
-    finality: {
-      blockNumber: 50_000_001n,
-      blockHash: `0x${'8'.repeat(64)}`,
-      blockTime: lateBoundary.block_time.toISOString(),
-      finalizedAt,
-    },
-    paymentResponseHeader: lateFinalityResponseHeader,
-  })
-  assert.equal(finalizedLatePayment.status, 'payment_pending')
-
-  const lateTiming = (await client.query<{
-    block_inside_authorization: boolean
-    block_inside_operation: boolean
-    observed_after_authorization: boolean
-    observed_after_operation: boolean
-    recovery_window_open: boolean
-  }>(`
-    SELECT finalized_block_time >= to_timestamp(x402_valid_after)
-        AND finalized_block_time < to_timestamp(x402_valid_before)
-        AS block_inside_authorization,
-      finalized_block_time >= start_time AND finalized_block_time < end_time
-        AS block_inside_operation,
-      finalized_at > to_timestamp(x402_valid_before) AS observed_after_authorization,
-      finalized_at > end_time AS observed_after_operation,
-      recovery_deadline_at > clock_timestamp() AS recovery_window_open
-    FROM payment_attempts WHERE public_id = $1
-  `, [lateFinalityAttemptId])).rows[0]
-  assert.deepEqual(lateTiming, {
-    block_inside_authorization: true,
-    block_inside_operation: true,
-    observed_after_authorization: true,
-    observed_after_operation: true,
-    recovery_window_open: true,
-  })
-
-  const lateCompleted = await completeTreasuryPaymentOperation(database, {
-    attemptId: lateFinalityAttemptId,
-    leaseOwner: lateFinalityLeaseOwner,
-  })
-  assert.equal(lateCompleted.state, 'completed')
-  if (lateCompleted.state !== 'completed') {
-    assert.fail('expected the late-finalizing drawing invention to complete inside recovery')
-  }
-  assert.equal(lateCompleted.method, 'x402')
-  assert.deepEqual(
-    (lateCompleted.response.kind as { drawing: unknown }).drawing,
-    lateFinalityDrawing,
+  await registerPaidRevisionTests(
+    t, client, database, placeId, drawing,
   )
-  const lateStored = (await client.query<{
-    status: string
-    payment_uses: number
-    kind_drawing: unknown
-  }>(`
-    SELECT attempt.status,
-      (SELECT count(*)::integer FROM payment_uses payment_use
-        WHERE payment_use.payment_attempt_id = attempt.public_id) AS payment_uses,
-      revision.drawing AS kind_drawing
-    FROM payment_attempts attempt
-    JOIN kinds kind ON kind.name = $2
-    JOIN kind_revisions revision
-      ON revision.kind_id = kind.id AND revision.revision = kind.current_revision
-    WHERE attempt.public_id = $1
-  `, [lateFinalityAttemptId, lateFinalityRequest.name])).rows[0]
-  assert.deepEqual(lateStored, {
-    status: 'completed',
-    payment_uses: 1,
-    kind_drawing: lateFinalityDrawing,
-  })
 
-  await issueCityFeeCredit(database, {
-    founderId: 1,
-    residentId: 2,
-    sourceKey: 'legacy-kind-invention-credit-0001',
-    reason: 'pre-drawing kind invention compatibility proof',
-  })
-  const legacyInvention = await beginCityCreditSpend(database, {
-    actorId: 2,
-    operation: 'kind_invention',
-    targetKey: 'kind-invention:legacy-undrawn-kind',
-    requestId: 'legacy-kind-invention-request-0001',
-    request: {
-      name: 'legacy-undrawn-kind',
-      description: 'the exact request shape stored before drawings existed',
-      traits: [],
-      recipe: [],
-    },
-  })
-  assert.equal(legacyInvention.state, 'ready')
-  if (legacyInvention.state !== 'ready') assert.fail('expected a ready legacy kind invention')
-  const legacyInvented = await completeTreasuryPaymentOperation(database, {
-    attemptId: legacyInvention.attempt_id,
-    leaseOwner: legacyInvention.lease_owner,
-  })
-  assert.equal(legacyInvented.state, 'completed')
-  if (legacyInvented.state !== 'completed') assert.fail('expected a completed legacy kind invention')
-  const legacyKind = legacyInvented.response.kind as { id: number; drawing: unknown }
-  assert.equal(legacyKind.drawing, null)
-  const legacyPinnedThingId = Number((await client.query<{ id: number }>(`
-    INSERT INTO things (
-      place_id, name, body, owner_id, maker_id, kind_id,
-      birth_revision, current_revision
-    ) VALUES ($1, 'legacy pinned thing', '', 2, 2, $2, 1, 1)
-    RETURNING id
-  `, [placeId, legacyKind.id])).rows[0]!.id)
+  await registerLateFinalityTests(
+    t, client, database, drawing,
+  )
 
-  await issueCityFeeCredit(database, {
-    founderId: 1,
-    residentId: 2,
-    sourceKey: 'legacy-kind-revision-credit-0001',
-    reason: 'pre-drawing kind revision compatibility proof',
-  })
-  const legacyRevision = await beginCityCreditSpend(database, {
-    actorId: 2,
-    operation: 'kind_revision',
-    targetKey: `kind-revision:${legacyKind.id}:2`,
-    requestId: 'legacy-kind-revision-request-0001',
-    assetType: 'kind',
-    assetId: legacyKind.id,
-    request: {
-      kind_id: legacyKind.id,
-      description: 'the old paid revision shape still settles',
-      traits: [],
-      recipe: [],
-    },
-  })
-  assert.equal(legacyRevision.state, 'ready')
-  if (legacyRevision.state !== 'ready') assert.fail('expected a ready legacy kind revision')
-  const legacyRevised = await completeTreasuryPaymentOperation(database, {
-    attemptId: legacyRevision.attempt_id,
-    leaseOwner: legacyRevision.lease_owner,
-  })
-  assert.equal(legacyRevised.state, 'completed')
-  if (legacyRevised.state !== 'completed') assert.fail('expected a completed legacy kind revision')
-  assert.equal((legacyRevised.response.kind as { drawing: unknown }).drawing, null)
-  assert.equal((await client.query<{ current_revision: number }>(`
-    SELECT current_revision FROM things WHERE id = $1
-  `, [legacyPinnedThingId])).rows[0]!.current_revision, 1)
+  await registerLegacyPaymentTests(
+    t, client, database, placeId,
+  )
 })
